+++ aidial_adapter_bedrock/llm/model/claude/adapter.py
@@ -12,15 +12,15 @@ from aidial_adapter_bedrock.bedrock import create_anthropic_client
 from aidial_adapter_bedrock.deployments import ChatCompletionDeployment as D
 from aidial_adapter_bedrock.deployments import ClaudeDeployment
 from aidial_adapter_bedrock.dial_api.storage import create_file_storage
-from aidial_adapter_bedrock.llm.converse.session_tags import (
-    resolve_session_tags,
-)
 from aidial_adapter_bedrock.llm.model.conf import CLAUDE_DEFAULT_MAX_TOKENS
 from aidial_adapter_bedrock.upstream_config import (
     CloudUpstreamConfig,
     UpstreamConfig,
 )
 from aidial_adapter_bedrock.utils.adapter_deployment import AdapterDeployment
+from aidial_adapter_bedrock.utils.session_tags import (
+    resolve_session_tags,
+)
 
 
 def _supports_thinking(deployment: ClaudeDeployment) -> bool:
